Add AsyncInputStream and FlowableFileStream

+++ client-runtime/src/main/java/com/microsoft/rest/v2/http/HttpRequestBody.java
@@ -17,7 +17,7 @@ public interface HttpRequestBody {
      * The length of this request body in bytes.
      * @return The length of this request body in bytes.
      */
-    int contentLength();
+    long contentLength();
 
     /**
      * @return the MIME Content-Type of this request body.
